TS: split RequestTasks and WorkflowTasks into distinct module
